[feature] Add API service capability
New methods on Wrapper interface:

Returns the address of the internal socket
	`Addr() net.Addr`

Return a new interface extending the `Wrapper` with API server
functionality. Use the returned object's `Listen` method to
start the API server using the default configuration.
	`ExposeAPI(ParseFunc) WrapperAPI`

New interface: `WrapperAPI`
	All `Wrapper` methods are directly accessible
	(extends the wrapper that created it)

Additional Methods:

Start an HTTP API server listening on the given address and path.
See the default request handler below for expected usage.
	`Listen(addr, path string) error`

Request handler used by `Listen`. Use this to integrate with
a larger API or setup a customized HTTP server listener. This
expects to receive an `HTTP POST` containing a JSON array of
strings. This will be interpreted as a command and arguments
to be sent to the wrapped process, with the lines of response
being transmitted to the client as another JSON string array.
	`CommandEndpoint(http.ResponseWriter, *http.Request)`

[feature] Add ParseFunc generation utilities

Added support for generating tables mapping command arugments to
their corresponding socketcmd headers, including subcommands.
See `examples/parsing.go` for more information.

Also added the standard header `-:`, corresponding to a command
that is not allowed to be run by the wrapped process. If this
header is used then the command will not be sent and an error
explaining why will be returned to the client. This allows
applications to set up command whitelists and/or blacklists.

// table maps commands to their headers
table := map[string]string{}

// create new `Argument` object from table
args := NewArguments(table, defaultHeader)

Methods:

Add the given argument table to the Argument spec
`args.AddArguments(table map[string]string)`

Add nested arguments to the Argument spec
`args.AddNestedArguments(table map[string]map[string]string)`

Export a new ParseFunc from the Argument spec
`args.ParseFunc()`